Bugfix: "movement" mode now handles correctly higher canvas sizes
Previously, the gap between two successive distances was constant, so with higher canvas sizes, one could see the jump from one distance to the next.
This is now fixed: the bigger the canvas/zoom, the smallest the distance gap is.
To do so, it is no longer the Parameters that compute the amount of points, it is now the job of the drawing handlers. The Parameters now only provides:
  - an indicator of how many points to draw (depending only on intensity and quality)
  - an indicator of the drawing size (depending only on canvas size and zoom)
Each handler uses these indicators accordingly to their own need.